loop_send works again
git-svn-id: file:///Users/walter/linux/walter/src/svnroot/trunk/bbs100@524 d90a1141-1c03-0410-82d0-813645769019

+++ src/Conn.c
@@ -227,12 +227,12 @@ void close_Conn(Conn *conn) {
 	conn->state = CONN_CLOSED;
 }
 
-void loop_Conn(Conn *conn, int iter) {
+void loop_Conn(Conn *conn, unsigned long iter) {
 	if (conn == NULL)
 		return;
 
 	conn->loop_counter = iter;
-	conn->state = (iter > 0) ? CONN_LOOPING : CONN_ESTABLISHED;
+	conn->state = (iter > 0UL) ? CONN_LOOPING : CONN_ESTABLISHED;
 }
 
 /* EOB */
