+++ src/include/Conn.h
@@ -94,7 +94,7 @@ int flush_Conn(Conn *);
 int input_Conn(Conn *);
 
 void close_Conn(Conn *);
-void loop_Conn(Conn *, int);
+void loop_Conn(Conn *, unsigned long);
 
 #endif	/* CONN_H_WJ105 */
